DEC-1438 - Spreadsheet cleanup
Convocate PO needs data clarity when users label paragraphs in crowdsourcing:

Note about changes made to spreadsheet:
Renamed actors to "feedback_topic" and added "feedback_topic_category" field to specify the top level category that the feedback topic belongs to. Also reorder columns a little to try to get the important stuff in the first few columns.

Added a second sheet to the Google Drive sheet with a description of the fields provided in the feedback sheet: